# sem_spring-rest/openapi.yaml
openapi: 3.0.0
info:
  title: Orders API
  version: v1
servers:
  - url: http://localhost:8080
    description: Local server
tags:
  - name: Orders
    description: CRUD operations for orders
paths:
  /api/orders:
    get:
      tags:
        - Orders
      summary: Get all orders
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/OrderStatus"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      tags:
        - Orders
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/orders/{id}:
    get:
      tags:
        - Orders
      summary: Get order by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    put:
      tags:
        - Orders
      summary: Update order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      tags:
        - Orders
      summary: Delete order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "204":
          description: Deleted
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
    OrderStatus:
      type: string
      enum:
        - NEW
        - CONFIRMED
        - CANCELLED
    OrderItemRequest:
      type: object
      properties:
        sku:
          type: string
          minLength: 3
          maxLength: 32
        quantity:
          type: integer
          format: int32
          minimum: 1
        price:
          type: number
          format: double
          minimum: 0.01
      required:
        - sku
        - quantity
        - price
    CreateOrderRequest:
      type: object
      properties:
        customerEmail:
          type: string
          format: email
          minLength: 1
        deliveryAddress:
          type: string
          minLength: 1
          maxLength: 200
        items:
          type: array
          minItems: 1
          maxItems: 20
          items:
            $ref: "#/components/schemas/OrderItemRequest"
      required:
        - customerEmail
        - deliveryAddress
        - items
    OrderItem:
      type: object
      properties:
        sku:
          type: string
          minLength: 3
          maxLength: 32
        quantity:
          type: integer
          format: int32
          minimum: 1
        price:
          type: number
          format: double
          minimum: 0.01
        lineTotal:
          type: number
          format: double
          minimum: 0.01
      required:
        - sku
        - quantity
        - price
        - lineTotal
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        customerEmail:
          type: string
          format: email
          minLength: 1
        deliveryAddress:
          type: string
          minLength: 1
          maxLength: 200
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        totalAmount:
          type: number
          format: double
          minimum: 0.01
        status:
          $ref: "#/components/schemas/OrderStatus"
      required:
        - id
        - customerEmail
        - deliveryAddress
        - items
        - totalAmount
        - status
    ApiError:
      type: object
      properties:
        message:
          type: string
        details:
          type: array
          items:
            type: string
      required:
        - message